requirements.txt
streamlit>=1.30.0
docker>=7.0.0
psutil>=5.9.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0